requirements.txt
pandas==1.3.0
Pillow==9.4.0
plotly==5.13.1
streamlit==1.20.0
openpyxl
numpy